+++ src/javascript/top_bar.js
@@ -221,18 +221,16 @@ const windows_buttons = `
   <button onclick="g_window.minimize(); " id="minimize" style=" height: auto;"><svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" style="isolation:isolate" viewBox="0 0 24 24" width="24" height="24"><rect x="7" y="11.5" width="10" height="0.8" transform="matrix(1,0,0,1,0,0)" fill="var(--titleBar-icons-color)"/></svg></button>
   <button onclick="g_window.maximize(); " id="maximize" style=" height: auto;"><svg width="24" height="24" viewBox="0 0 50 50" fill="none" xmlns="http://www.w3.org/2000/svg"><rect x="16" y="16" width="18.5714" height="18.5714" stroke="var(--titleBar-icons-color)" stroke-width="2"/></svg></button>
   <button onclick="g_window.close();" id="close" style=" height: auto;"><svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" style="isolation:isolate" viewBox="0 0 24 24" width="20" height="24"><rect x="3.68" y="11.406" width="16.64" height="1.189" transform="matrix(-0.707107,0.707107,-0.707107,-0.707107,28.970563,12)" fill="var(--titleBar-icons-color)" /><rect x="3.68" y="11.406" width="16.64" height="1.189" transform="matrix(-0.707107,-0.707107,0.707107,-0.707107,12,28.970563)" fill="var(--titleBar-icons-color)" /></svg></button>
-`; 
+`;
 
 if (graviton.currentOS().codename == "win32") {
   document.getElementById("controls").innerHTML = windows_buttons;
   g_window.on("maximize", (e, cmd) => {
-    const button = document
-    .getElementById("maximize");
+    const button = document.getElementById("maximize");
     button.setAttribute("onclick", "g_window.unmaximize();");
   });
   g_window.on("unmaximize", (e, cmd) => {
-    const button = document
-    .getElementById("maximize");
+    const button = document.getElementById("maximize");
     button.setAttribute("onclick", "g_window.maximize();");
   });
 } else {